Feat 댓글의 답글

+++ src/main/java/project/ffboard/service/CommentService.java
@@ -5,7 +5,7 @@ import project.ffboard.dto.Comment;
 import java.util.List;
 
 public interface CommentService {
-    public int addComment(Comment comment);
+    public Long addComment(Comment comment);
     public int deleteComment(Long id);
     public int updateComment(Comment comment);
     public List<Comment> getCommentList(Long articleId);
